async_api/src/requirements.txt
aioelasticsearch==0.7.0
aioredis==1.3.1
fastapi==0.62.0
uvicorn==0.13.1
orjson==3.4.6
